+++ modules/module-billing-provisioning-livewire/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Added Livewire provisioned-service creation and lifecycle transitions.
+
 ## 0.1.0
 
 - Initial Provisioning Livewire 4 operation component.